requirements.txt
mediapipe==0.10.21
opencv-contrib-python
numpy
pyfluidsynth
pywebview
python-dotenv
pytest